% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_norm.R
\name{spp_norm}
\alias{spp_norm}
\title{Padroniza nomes científicos}
\usage{
spp_norm(x)
}
\arguments{
\item{x}{Vetor de caracteres contendo nomes científicos.}
}
\value{
Um vetor de caracteres, do mesmo tamanho de \code{x}, com os nomes
padronizados. Valores \code{NA} são mantidos como \code{NA}.
}
\description{
Normaliza nomes científicos removendo espaços extras,
convertendo todo o texto para minúsculo e aplicando
capitalização correta ao gênero.
}
\details{
A função:
\itemize{
\item Remove espaços no início e no fim;
\item Substitui múltiplos espaços por um único espaço;
\item Converte o texto para minúsculo;
\item Capitaliza apenas a primeira letra do gênero.
}
}
\examples{
spp_norm("   ACTINOPUS   ANSELMOI ")
# "Actinopus anselmoi"

spp_norm(c("LOXOSCELES intermedia", "phoneutria  nigriventer"))
# "Loxosceles intermedia" "Phoneutria nigriventer"

}
