% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider_family.R
\name{spider_family}
\alias{spider_family}
\title{Obtém a família taxonômica de uma espécie de aranha.}
\usage{
spider_family(especie, lsid = NA_character_)
}
\arguments{
\item{especie}{name da espécie (ex: "Actinopus anselmoi").}
}
\value{
name da família ou \code{NA_character_} quando não encontrado.
}
\description{
Obtém a família taxonômica de uma espécie de aranha.
}
